CHE-5718 Progress of the importing project is not displayed
